+++ tests/test_high_consumers.rs
@@ -1,13 +1,13 @@
 //! Test data consumption using high level consumers.
 
 use std::collections::HashMap;
-use std::sync::atomic::AtomicUsize;
+use std::error::Error;
 use std::sync::Arc;
-use std::time::Duration;
 
-use futures::future;
+use futures::future::{self, FutureExt};
 use futures::stream::StreamExt;
 use maplit::hashmap;
+use tokio::time::{self, Duration};
 
 use rdkafka::consumer::{CommitMode, Consumer, ConsumerContext, StreamConsumer};
 use rdkafka::error::KafkaError;
@@ -25,10 +25,7 @@ fn create_stream_consumer(
     group_id: &str,
     config_overrides: Option<HashMap<&str, &str>>,
 ) -> StreamConsumer<ConsumerTestContext> {
-    let cons_context = ConsumerTestContext {
-        _n: 64,
-        wakeups: Arc::new(AtomicUsize::new(0)),
-    };
+    let cons_context = ConsumerTestContext { _n: 64 };
     create_stream_consumer_with_context(group_id, config_overrides, cons_context)
 }
 
@@ -371,4 +368,157 @@ async fn test_consumer_store_offset_commit() {
         .add_partition_offset(&topic_name, 2, Offset::Offset(12))
         .unwrap();
     assert_eq!(position, consumer.position().unwrap());
+}
+
+#[tokio::test(flavor = "multi_thread")]
+async fn test_consumer_commit_metadata() -> Result<(), Box<dyn Error>> {
+    let _ = env_logger::try_init();
+
+    let topic_name = rand_test_topic();
+    let group_name = rand_test_group();
+    populate_topic(&topic_name, 10, &value_fn, &key_fn, None, None).await;
+
+    let create_consumer = || async {
+        // Disable auto-commit so we can manually drive the commits.
+        let mut config = HashMap::new();
+        config.insert("enable.auto.commit", "false");
+        let consumer = create_stream_consumer(&group_name, Some(config));
+
+        // Subscribe to the topic and wait for at least one message, which
+        // ensures that the consumer group has been joined and such.
+        consumer.subscribe(&[topic_name.as_str()])?;
+        let _ = consumer.stream().next().await;
+
+        Ok::<_, Box<dyn Error>>(consumer)
+    };
+
+    // Create a topic partition list where each element has some associated
+    // metadata.
+    let tpl = {
+        let mut tpl = TopicPartitionList::new();
+        let mut tpl1 = tpl.add_partition(&topic_name, 0);
+        tpl1.set_offset(Offset::Offset(1))?;
+        tpl1.set_metadata("one");
+        let mut tpl2 = tpl.add_partition(&topic_name, 1);
+        tpl2.set_offset(Offset::Offset(1))?;
+        tpl2.set_metadata("two");
+        let mut tpl3 = tpl.add_partition(&topic_name, 2);
+        tpl3.set_offset(Offset::Offset(1))?;
+        tpl3.set_metadata("three");
+        tpl
+    };
+
+    // Ensure that the commit state immediately includes the metadata.
+    {
+        let consumer = create_consumer().await?;
+        consumer.commit(&tpl, CommitMode::Sync)?;
+        assert_eq!(consumer.committed(None)?, tpl);
+    }
+
+    // Ensure that the commit state on a new consumer in the same group
+    // can see the same metadata.
+    {
+        let consumer = create_consumer().await?;
+        assert_eq!(consumer.committed(None)?, tpl);
+    }
+
+    Ok(())
+}
+
+#[tokio::test]
+async fn test_consume_partition_order() {
+    let _r = env_logger::try_init();
+
+    let topic_name = rand_test_topic();
+    populate_topic(&topic_name, 4, &value_fn, &key_fn, Some(0), None).await;
+    populate_topic(&topic_name, 4, &value_fn, &key_fn, Some(1), None).await;
+    populate_topic(&topic_name, 4, &value_fn, &key_fn, Some(2), None).await;
+
+    // Using partition queues should allow us to consume the partitions
+    // in a round-robin fashion.
+    {
+        let consumer = Arc::new(create_stream_consumer(&rand_test_group(), None));
+        let mut tpl = TopicPartitionList::new();
+        tpl.add_partition_offset(&topic_name, 0, Offset::Beginning)
+            .unwrap();
+        tpl.add_partition_offset(&topic_name, 1, Offset::Beginning)
+            .unwrap();
+        tpl.add_partition_offset(&topic_name, 2, Offset::Beginning)
+            .unwrap();
+        consumer.assign(&tpl).unwrap();
+
+        let mut partition_streams: Vec<_> = (0..3)
+            .map(|i| consumer.split_partition_queue(&topic_name, i).unwrap())
+            .collect();
+
+        for _ in 0..4 {
+            let main_message =
+                time::timeout(Duration::from_millis(100), consumer.stream().next()).await;
+            assert!(main_message.is_err());
+
+            for (i, stream) in partition_streams.iter_mut().enumerate() {
+                let queue_message = stream.recv().await.unwrap();
+                assert_eq!(queue_message.partition(), i as i32);
+            }
+        }
+    }
+
+    // When not all partitions have been split into separate queues, the
+    // unsplit partitions should still be accessible via the main queue.
+    {
+        let consumer = Arc::new(create_stream_consumer(&rand_test_group(), None));
+        let mut tpl = TopicPartitionList::new();
+        tpl.add_partition_offset(&topic_name, 0, Offset::Beginning)
+            .unwrap();
+        tpl.add_partition_offset(&topic_name, 1, Offset::Beginning)
+            .unwrap();
+        tpl.add_partition_offset(&topic_name, 2, Offset::Beginning)
+            .unwrap();
+        consumer.assign(&tpl).unwrap();
+
+        let partition1 = consumer.split_partition_queue(&topic_name, 1).unwrap();
+
+        let mut i = 0;
+        while i < 12 {
+            if let Some(m) = consumer.recv().now_or_never() {
+                let partition = m.unwrap().partition();
+                assert!(partition == 0 || partition == 2);
+                i += 1;
+            }
+
+            if let Some(m) = partition1.recv().now_or_never() {
+                assert_eq!(m.unwrap().partition(), 1);
+                i += 1;
+            }
+        }
+    }
+
+    // Sending the queue to another task that is likely to outlive the
+    // original thread should work. This is not idiomatic, as the consumer
+    // should be continuously polled to serve callbacks, but it should not panic
+    // or result in memory unsafety, etc.
+    {
+        let consumer = Arc::new(create_stream_consumer(&rand_test_group(), None));
+        let mut tpl = TopicPartitionList::new();
+        tpl.add_partition_offset(&topic_name, 0, Offset::Beginning)
+            .unwrap();
+        consumer.assign(&tpl).unwrap();
+        let stream = consumer.split_partition_queue(&topic_name, 0).unwrap();
+
+        let worker = tokio::spawn({
+            async move {
+                for _ in 0..4 {
+                    let stream_message = stream.recv().await.unwrap();
+                    assert_eq!(stream_message.partition(), 0);
+                }
+            }
+        });
+
+        let main_message =
+            time::timeout(Duration::from_millis(100), consumer.stream().next()).await;
+        assert!(main_message.is_err());
+
+        drop(consumer);
+        worker.await.unwrap();
+    }
 }